feat(design-system): Initialize design tokens (#300)
Closes #37, #38, #39.
This initializes our design token system, which we use to sync design decisions across Vue, Figma, Canvas, etc. The design system has a `design-system-tokens` package which stores the tokens and a `design-system-token-transformer` package which scripts the translation from JSON design tokens to CSS variables. The tokens are not yet complete; they will be expanded as we build out our components further.