Integrate Inkline component library as baseline
We plan to integrate the Inkline Vue component library as a starting
point to speed component creation. We will wrap Inkline components in
an Eterna components and use the Eterna components in our pages. This
gives us an escape hatch should we need to replace the Inkline library
in the future, and lets us define Eterna-specific details for each
component.